Migrate to the idiomatic Prettier recursion.
Changing this enables progressive migration of node printers to the
idiomatic Prettier recursion using the generic `print` function and
the `FastPath` stack. More importantly, it is a first step to
support comments.

To do so, the ANTLR parse tree is preprocessed to a simplified
version that is compatible with the `FastPath` stack accessors.

ANTLR doesn't distinguish between `identifier?` and `identifier*`
and stores all children of a parse node in lists. When a single
child is expected, `path.call` is therefore called with an awkward,
trailing argument `0` (equivalent to `list[0]`).